unit error parsing + F to C order iltm
Former-commit-id: e21f4ced36cbb7cc4118e6cf502d28166f5926ff [formerly 8ac52c054287903a37732c3adc3affca071f9447]
Former-commit-id: ce9415155e05c22dbbdef645e11d85dfd2da91ee

+++ dtapy/core/jitclasses.py
@@ -327,9 +327,9 @@ class DynamicDemand(object):
         # time step traffic is loaded into the network
         self.all_destinations = get_all_destinations(demands)  # for destination/origin based labels
         self.all_origins = get_all_origins(demands)
-        self.tot_time_steps = tot_time_steps
-        self.tot_origins = self.all_destinations.size
-        self.tot_destinations = self.all_destinations.size
+        self.tot_time_steps = np.uint32(tot_time_steps)
+        self.tot_origins = np.uint32(self.all_destinations.size)
+        self.tot_destinations = np.uint32(self.all_destinations.size)
 
     def update(self):
         self.__time_step += uint32(1)
